+++ src/services/spotifyplus-service.ts
@@ -211,8 +211,12 @@ export class SpotifyPlusService {
       }
 
       // validations.
-      if (device_id == null)
-        device_id = player.attributes.sp_device_id || player.attributes.source || null;
+      if (device_id == null) {
+        if (player.attributes.sp_device_is_brand_sonos)
+          device_id = player.attributes.source || null;
+        else
+          device_id = player.attributes.sp_device_id || player.attributes.source || null;
+      }
 
       // if default device configured then override the specified deviceId.
       if (this.config.deviceDefaultId) {
@@ -2967,8 +2971,12 @@ export class SpotifyPlusService {
       // validations.
       if (!context_uri)
         throw new Error("STPC0005 context_uri argument was not supplied to the PlayerMediaPlayContext service.")
-      if (device_id == null)
-        device_id = player.attributes.sp_device_id || player.attributes.source || null;
+      if (device_id == null) {
+        if (player.attributes.sp_device_is_brand_sonos)
+          device_id = player.attributes.source || null;
+        else
+          device_id = player.attributes.sp_device_id || player.attributes.source || null;
+      }
 
       // if default device configured then override the specified deviceId.
       if (this.config.deviceDefaultId) {
@@ -3054,8 +3062,12 @@ export class SpotifyPlusService {
       }
 
       // validations.
-      if (device_id == null)
-        device_id = player.attributes.sp_device_id || player.attributes.source || null;
+      if (device_id == null) {
+        if (player.attributes.sp_device_is_brand_sonos)
+          device_id = player.attributes.source || null;
+        else
+          device_id = player.attributes.sp_device_id || player.attributes.source || null;
+      }
 
       // if default device configured then override the specified deviceId.
       if (this.config.deviceDefaultId) {
@@ -3143,8 +3155,12 @@ export class SpotifyPlusService {
         throw new Error("STPC0005 uris argument was not supplied to the PlayerMediaPlayTracks service.")
       if (position_ms == null)
         position_ms = 0;
-      if (device_id == null)
-        device_id = player.attributes.sp_device_id || player.attributes.source || null;
+      if (device_id == null) {
+        if (player.attributes.sp_device_is_brand_sonos)
+          device_id = player.attributes.source || null;
+        else
+          device_id = player.attributes.sp_device_id || player.attributes.source || null;
+      }
 
       // if default device configured then override the specified deviceId.
       if (this.config.deviceDefaultId) {
@@ -3216,8 +3232,12 @@ export class SpotifyPlusService {
       }
 
       // validations.
-      if (device_id == null)
-        device_id = player.attributes.sp_device_id || player.attributes.source || null;
+      if (device_id == null) {
+        if (player.attributes.sp_device_is_brand_sonos)
+          device_id = player.attributes.source || null;
+        else
+          device_id = player.attributes.sp_device_id || player.attributes.source || null;
+      }
 
       // if default device configured then override the specified deviceId.
       if (this.config.deviceDefaultId) {
@@ -3314,8 +3334,12 @@ export class SpotifyPlusService {
       }
 
       // validations.
-      if (device_id == null)
-        device_id = player.attributes.sp_device_id || player.attributes.source || null;
+      if (device_id == null) {
+        if (player.attributes.sp_device_is_brand_sonos)
+          device_id = player.attributes.source || null;
+        else
+          device_id = player.attributes.sp_device_id || player.attributes.source || null;
+      }
       if (play == null)
         play = true;